+++ pyogrio/tests/test_geopandas_io.py
@@ -7,8 +7,16 @@ from zipfile import ZipFile
 
 import numpy as np
 
-from pyogrio import __gdal_version__, list_drivers, list_layers, read_info
-from pyogrio._compat import HAS_ARROW_WRITE_API, HAS_PYPROJ, PANDAS_GE_15
+from pyogrio import (
+    __gdal_version__,
+    list_drivers,
+    list_layers,
+    read_info,
+    set_gdal_config_options,
+    vsi_listtree,
+    vsi_unlink,
+)
+from pyogrio._compat import GDAL_GE_352, HAS_ARROW_WRITE_API, HAS_PYPROJ, PANDAS_GE_15
 from pyogrio.errors import DataLayerError, DataSourceError, FeatureError, GeometryError
 from pyogrio.geopandas import PANDAS_GE_20, read_dataframe, write_dataframe
 from pyogrio.raw import (
@@ -18,6 +26,7 @@ from pyogrio.raw import (
 from pyogrio.tests.conftest import (
     ALL_EXTS,
     DRIVERS,
+    START_FID,
     requires_arrow_write_api,
     requires_gdal_geos,
     requires_pyarrow_api,
@@ -219,6 +228,22 @@ def test_read_force_2d(tmp_path, use_arrow):
     assert not df.iloc[0].geometry.has_z
 
 
+@pytest.mark.skipif(
+    not GDAL_GE_352,
+    reason="gdal >= 3.5.2 needed to use OGR_GEOJSON_MAX_OBJ_SIZE with a float value",
+)
+def test_read_geojson_error(naturalearth_lowres_geojson, use_arrow):
+    try:
+        set_gdal_config_options({"OGR_GEOJSON_MAX_OBJ_SIZE": 0.01})
+        with pytest.raises(
+            DataSourceError,
+            match="Failed to read GeoJSON data; .* GeoJSON object too complex",
+        ):
+            read_dataframe(naturalearth_lowres_geojson, use_arrow=use_arrow)
+    finally:
+        set_gdal_config_options({"OGR_GEOJSON_MAX_OBJ_SIZE": None})
+
+
 def test_read_layer(tmp_path, use_arrow):
     filename = tmp_path / "test.gpkg"
 
@@ -371,12 +396,9 @@ def test_read_fid_as_index(naturalearth_lowres_all_ext, use_arrow):
         fid_as_index=True,
         **kwargs,
     )
-    if naturalearth_lowres_all_ext.suffix in [".gpkg"]:
-        # File format where fid starts at 1
-        assert_index_equal(df.index, pd.Index([3, 4], name="fid"))
-    else:
-        # File format where fid starts at 0
-        assert_index_equal(df.index, pd.Index([2, 3], name="fid"))
+    fids_expected = pd.Index([2, 3], name="fid")
+    fids_expected += START_FID[naturalearth_lowres_all_ext.suffix]
+    assert_index_equal(df.index, fids_expected)
 
 
 def test_read_fid_as_index_only(naturalearth_lowres, use_arrow):
@@ -1568,6 +1590,22 @@ def test_write_read_null(tmp_path, use_arrow):
     assert result_gdf["object_str"][2] is None
 
 
+@pytest.mark.requires_arrow_write_api
+def test_write_read_vsimem(naturalearth_lowres_vsi, use_arrow):
+    path, _ = naturalearth_lowres_vsi
+    mem_path = f"/vsimem/{path.name}"
+
+    input = read_dataframe(path, use_arrow=use_arrow)
+    assert len(input) == 177
+
+    try:
+        write_dataframe(input, mem_path, use_arrow=use_arrow)
+        result = read_dataframe(mem_path, use_arrow=use_arrow)
+        assert len(result) == 177
+    finally:
+        vsi_unlink(mem_path)
+
+
 @pytest.mark.parametrize(
     "wkt,geom_types",
     [
@@ -1744,6 +1782,10 @@ def test_read_invalid_poly_ring(tmp_path, use_arrow, on_invalid, message):
 
 def test_read_multisurface(multisurface_file, use_arrow):
     if use_arrow:
+        # TODO: revisit once https://github.com/geopandas/pyogrio/issues/478
+        # is resolved.
+        pytest.skip("Shapely + GEOS 3.13 crashes in from_wkb for this case")
+
         with pytest.raises(shapely.errors.GEOSException):
             # TODO(Arrow)
             # shapely fails parsing the WKB
@@ -1974,6 +2016,9 @@ def test_write_memory(naturalearth_lowres, driver):
         check_dtype=not is_json,
     )
 
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
+
 
 def test_write_memory_driver_required(naturalearth_lowres):
     df = read_dataframe(naturalearth_lowres)
@@ -1985,6 +2030,9 @@ def test_write_memory_driver_required(naturalearth_lowres):
         match="driver must be provided to write to in-memory file",
     ):
         write_dataframe(df.head(1), buffer, driver=None, layer="test")
+
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
 
 
 @pytest.mark.parametrize("driver", ["ESRI Shapefile", "OpenFileGDB"])
@@ -2001,6 +2049,9 @@ def test_write_memory_unsupported_driver(naturalearth_lowres, driver):
     ):
         write_dataframe(df, buffer, driver=driver, layer="test")
 
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
+
 
 @pytest.mark.parametrize("driver", ["GeoJSON", "GPKG"])
 def test_write_memory_append_unsupported(naturalearth_lowres, driver):
@@ -2013,6 +2064,9 @@ def test_write_memory_append_unsupported(naturalearth_lowres, driver):
     ):
         write_dataframe(df.head(1), buffer, driver=driver, layer="test", append=True)
 
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
+
 
 def test_write_memory_existing_unsupported(naturalearth_lowres):
     df = read_dataframe(naturalearth_lowres)
@@ -2023,6 +2077,9 @@ def test_write_memory_existing_unsupported(naturalearth_lowres):
         match="writing to existing in-memory object is not supported",
     ):
         write_dataframe(df.head(1), buffer, driver="GeoJSON", layer="test")
+
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
 
 
 def test_write_open_file_handle(tmp_path, naturalearth_lowres):
@@ -2044,6 +2101,9 @@ def test_write_open_file_handle(tmp_path, naturalearth_lowres):
         with ZipFile(tmp_path / "test.geojson.zip", "w") as z:
             with z.open("test.geojson", "w") as f:
                 write_dataframe(df.head(1), f)
+
+    # Check temp file was cleaned up. Filter, as gdal keeps cache files in /vsimem/.
+    assert vsi_listtree("/vsimem/", pattern="pyogrio_*") == []
 
 
 @pytest.mark.parametrize("ext", ["gpkg", "geojson"])
@@ -2214,7 +2274,7 @@ def test_write_kml_file_coordinate_order(tmp_path, use_arrow):
     if "LIBKML" in list_drivers():
         # test appending to the existing file only if LIBKML is available
         # as it appears to fall back on LIBKML driver when appending.
-        points_append = [Point(70, 80), Point(90, 100), Point(110, 120)]
+        points_append = [Point(7, 8), Point(9, 10), Point(11, 12)]
         gdf_append = gp.GeoDataFrame(geometry=points_append, crs="EPSG:4326")
 
         write_dataframe(